Separate JobExplorerListRow into a new js file
